refactored to avoid overuse of the word commands
